+++ modules/index.js
@@ -1,7 +1,7 @@
-export {IOProvider} from './context'
+export {IOProvider, Context as IOContext} from './context'
 import {withIO} from './withIO'
 
 export {withIO}
 export const withObservables = withIO
 
-export {useIO} from './useIO'
+export {useIO, pruneCache} from './useIO'