+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 5.2.0 - 2020-11-12
+
+### Added
+
+- IOContext export.
+- pruneCache function to clean orphaned cache entries.
+
+### Fixed
+
+- Really fix useIO initial error handling. Adds request cache outside of react state.
+
 ## 5.1.2 - 2020-11-07
 
 ### Fixed